chore(ci): gate feature-matrix clippy and fix pre-existing warnings
CI only ran `-D warnings` clippy on the default feature set; the slim-default +
optional-features design means feature-gated code was never lint-gated. Add a
feature-matrix clippy step and fix the warnings it surfaced:

- repository.rs: drop 5 no-op quote-escaping calls that replaced `"` with `"`
  (a no-op); the SQL single-quote escaping (`'` -> `''`) is correct and unchanged.
- connector.rs: redundant closure -> `DashMap::new`.
- data.rs: drop redundant `as usize` casts (`height()`/`width()` already return usize).

+++ src/common/model/data.rs
@@ -327,7 +327,7 @@ impl DataEvent {
             #[cfg(feature = "polars")]
             DataType::DataFrame(df_store) => match &df_store.data {
                 DataframeStoreData::Bytes(bytes) => bytes.len(),
-                DataframeStoreData::DataFrame(df) => df.height() as usize * df.width() as usize, // rough estimate: rows * columns
+                DataframeStoreData::DataFrame(df) => df.height() * df.width(), // rough estimate: rows * columns
             },
             DataType::File(file_store) => file_store.content.len(),
         }
